+++ httprunner/client.py
@@ -4,11 +4,12 @@ import time
 
 import requests
 import urllib3
-from httprunner import logger
-from httprunner.utils import lower_dict_keys, omit_long_data
 from requests import Request, Response
 from requests.exceptions import (InvalidSchema, InvalidURL, MissingSchema,
                                  RequestException)
+
+from httprunner import logger, response
+from httprunner.utils import lower_dict_keys, omit_long_data
 
 urllib3.disable_warnings(urllib3.exceptions.InsecureRequestWarning)
 
@@ -30,8 +31,8 @@ class HttpSession(requests.Session):
     This is a slightly extended version of `python-request <http://python-requests.org>`_'s
     :py:class:`requests.Session` class and mostly this class works exactly the same.
     """
-    def __init__(self, *args, **kwargs):
-        super(HttpSession, self).__init__(*args, **kwargs)
+    def __init__(self):
+        super(HttpSession, self).__init__()
         self.init_meta_data()
 
     def init_meta_data(self):
@@ -114,7 +115,10 @@ class HttpSession(requests.Session):
         else:
             try:
                 # try to record json data
-                req_resp_dict["response"]["json"] = resp_obj.json()
+                if isinstance(resp_obj, response.ResponseObject):
+                    req_resp_dict["response"]["json"] = resp_obj.json
+                else:
+                    req_resp_dict["response"]["json"] = resp_obj.json()
             except ValueError:
                 # only record at most 512 text charactors
                 resp_text = resp_obj.text
@@ -124,6 +128,13 @@ class HttpSession(requests.Session):
         log_print(req_resp_dict, "response")
 
         return req_resp_dict
+
+    def update_last_req_resp_record(self, resp_obj):
+        """
+        update request and response info from Response() object.
+        """
+        self.meta_data["data"].pop()
+        self.meta_data["data"].append(self.get_req_resp_record(resp_obj))
 
     def request(self, method, url, name=None, **kwargs):
         """