+++ httprunner/plugins/locusts/__main__.py
@@ -1,0 +1,4 @@
+from httprunner.plugins.locusts.cli import main
+
+if __name__ == "__main__":
+    main()