+++ tests/test_api.py
@@ -2,10 +2,9 @@ import os
 import re
 import shutil
 import time
-import unittest
 
-from httprunner import exceptions, loader, parser
-from httprunner.api import HttpRunner, prepare_locust_tests
+from httprunner import exceptions, loader, parser, report
+from httprunner.api import HttpRunner
 from tests.api_server import HTTPBIN_SERVER
 from tests.base import ApiServerUnittest
 
@@ -35,7 +34,8 @@ class TestHttpRunner(ApiServerUnittest):
                     'request': {
                         'url': 'http://127.0.0.1:5000/api/get-token',
                         'method': 'POST',
-                        'headers': {'Content-Type': 'application/json', 'app_version': '2.8.6', 'device_sn': 'FwgRiO7CNA50DSU', 'os_platform': 'ios', 'user_agent': 'iOS/10.3'},
+                        'headers': {'Content-Type': 'application/json', 'app_version': '2.8.6',
+                                    'device_sn': 'FwgRiO7CNA50DSU', 'os_platform': 'ios', 'user_agent': 'iOS/10.3'},
                         'json': {'sign': '9c0c7e51c91ae963c833a4ccbab8d683c4a90c98'}
                     },
                     'extract': [
@@ -52,7 +52,8 @@ class TestHttpRunner(ApiServerUnittest):
                     'request': {
                         'url': 'http://127.0.0.1:5000/api/users/1000',
                         'method': 'POST',
-                        'headers': {'Content-Type': 'application/json', 'device_sn': 'FwgRiO7CNA50DSU','token': '$token'},
+                        'headers': {'Content-Type': 'application/json',
+                                    'device_sn': 'FwgRiO7CNA50DSU','token': '$token'},
                         'json': {'name': 'user1', 'password': '123456'}
                     },
                     'validate': [
@@ -76,9 +77,9 @@ class TestHttpRunner(ApiServerUnittest):
         return self.api_client.get(url, headers=headers)
 
     def test_text_run_times(self):
-        self.runner.run(self.testcase_cli_path)
-        self.assertEqual(self.runner.summary["stat"]["testcases"]["total"], 1)
-        self.assertEqual(self.runner.summary["stat"]["teststeps"]["total"], 10)
+        summary = self.runner.run(self.testcase_cli_path)
+        self.assertEqual(summary["stat"]["testcases"]["total"], 1)
+        self.assertEqual(summary["stat"]["teststeps"]["total"], 10)
 
     def test_text_run_times_invalid(self):
         testcases = [
@@ -114,8 +115,8 @@ class TestHttpRunner(ApiServerUnittest):
             self.runner.run_tests(tests_mapping)
 
     def test_text_skip(self):
-        self.runner.run(self.testcase_cli_path)
-        self.assertEqual(self.runner.summary["stat"]["teststeps"]["skipped"], 4)
+        summary = self.runner.run(self.testcase_cli_path)
+        self.assertEqual(summary["stat"]["teststeps"]["skipped"], 4)
 
     def test_save_variables_output(self):
         testcases = [
@@ -180,27 +181,29 @@ class TestHttpRunner(ApiServerUnittest):
         self.assertEqual(token1, token2)
 
     def test_html_report(self):
-        report_save_dir = os.path.join(os.getcwd(), 'reports', "demo")
-        runner = HttpRunner(failfast=True, report_dir=report_save_dir)
-        runner.run(self.testcase_cli_path)
-        summary = runner.summary
+        runner = HttpRunner(failfast=True)
+        summary = runner.run(self.testcase_cli_path)
         self.assertEqual(summary["stat"]["testcases"]["total"], 1)
         self.assertEqual(summary["stat"]["teststeps"]["total"], 10)
         self.assertEqual(summary["stat"]["teststeps"]["skipped"], 4)
+
+        report_save_dir = os.path.join(os.getcwd(), 'reports', "demo")
+        report.gen_html_report(summary, report_dir=report_save_dir)
         self.assertGreater(len(os.listdir(report_save_dir)), 0)
         shutil.rmtree(report_save_dir)
 
     def test_html_report_with_fixed_report_file(self):
-        report_save_dir = os.path.join(os.getcwd(), 'reports', "demo")
-        report_file = 'test.html'
-        runner = HttpRunner(failfast=True, report_dir=report_save_dir, report_file=report_file)
-        runner.run(self.testcase_cli_path)
-        summary = runner.summary
+        runner = HttpRunner(failfast=True)
+        summary = runner.run(self.testcase_cli_path)
         self.assertEqual(summary["stat"]["testcases"]["total"], 1)
         self.assertEqual(summary["stat"]["teststeps"]["total"], 10)
         self.assertEqual(summary["stat"]["teststeps"]["skipped"], 4)
+
+        report_file = os.path.join(os.getcwd(), 'reports', "demo", "test.html")
+        report.gen_html_report(summary, report_file=report_file)
+        report_save_dir = os.path.dirname(report_file)
         self.assertEqual(len(os.listdir(report_save_dir)), 1)
-        self.assertTrue(os.path.isfile(os.path.join(report_save_dir, report_file)))
+        self.assertTrue(os.path.isfile(report_file))
         shutil.rmtree(report_save_dir)
 
     def test_log_file(self):
@@ -211,8 +214,7 @@ class TestHttpRunner(ApiServerUnittest):
         os.remove(log_file_path)
 
     def test_run_testcases(self):
-        self.runner.run_tests(self.tests_mapping)
-        summary = self.runner.summary
+        summary = self.runner.run_tests(self.tests_mapping)
         self.assertTrue(summary["success"])
         self.assertEqual(summary["stat"]["testcases"]["total"], 1)
         self.assertEqual(summary["stat"]["teststeps"]["total"], 2)
@@ -220,8 +222,7 @@ class TestHttpRunner(ApiServerUnittest):
         self.assertIn("records", summary["details"][0])
 
     def test_run_yaml_upload(self):
-        self.runner.run("tests/httpbin/upload.yml")
-        summary = self.runner.summary
+        summary = self.runner.run("tests/httpbin/upload.yml")
         self.assertTrue(summary["success"])
         self.assertEqual(summary["stat"]["testcases"]["total"], 1)
         self.assertEqual(summary["stat"]["teststeps"]["total"], 1)
@@ -257,8 +258,7 @@ class TestHttpRunner(ApiServerUnittest):
         tests_mapping = {
             "testcases": testcases
         }
-        self.runner.run_tests(tests_mapping)
-        summary = self.runner.summary
+        summary = self.runner.run_tests(tests_mapping)
         self.assertTrue(summary["success"])
         self.assertEqual(summary["stat"]["testcases"]["total"], 1)
         self.assertEqual(summary["stat"]["teststeps"]["total"], 1)
@@ -268,23 +268,23 @@ class TestHttpRunner(ApiServerUnittest):
         )
 
     def test_html_report_repsonse_image(self):
+        runner = HttpRunner(failfast=True)
+        summary = runner.run("tests/httpbin/load_image.yml")
+
         report_save_dir = os.path.join(os.getcwd(), 'reports', "demo")
-        runner = HttpRunner(failfast=True, report_dir=report_save_dir)
-        report = runner.run("tests/httpbin/load_image.yml")
-        self.assertTrue(os.path.isfile(report))
+        report_path = report.gen_html_report(summary, report_dir=report_save_dir)
+        self.assertTrue(os.path.isfile(report_path))
         shutil.rmtree(report_save_dir)
 
     def test_testcase_layer_with_api(self):
-        self.runner.run("tests/testcases/setup.yml")
-        summary = self.runner.summary
+        summary = self.runner.run("tests/testcases/setup.yml")
         self.assertTrue(summary["success"])
         self.assertEqual(summary["details"][0]["records"][0]["name"], "get token (setup)")
         self.assertEqual(summary["stat"]["testcases"]["total"], 1)
         self.assertEqual(summary["stat"]["teststeps"]["total"], 2)
 
     def test_testcase_layer_with_testcase(self):
-        self.runner.run("tests/testsuites/create_users.yml")
-        summary = self.runner.summary
+        summary = self.runner.run("tests/testsuites/create_users.yml")
         self.assertTrue(summary["success"])
         self.assertEqual(summary["stat"]["testcases"]["total"], 2)
         self.assertEqual(summary["stat"]["teststeps"]["total"], 4)
@@ -293,9 +293,8 @@ class TestHttpRunner(ApiServerUnittest):
         testcase_file_path = os.path.join(
             os.getcwd(), 'tests/httpbin/hooks.yml')
         start_time = time.time()
-        self.runner.run(testcase_file_path)
+        summary = self.runner.run(testcase_file_path)
         end_time = time.time()
-        summary = self.runner.summary
         self.assertTrue(summary["success"])
         self.assertLess(end_time - start_time, 60)
 
@@ -333,8 +332,7 @@ class TestHttpRunner(ApiServerUnittest):
             "project_mapping": loader.project_mapping,
             "testcases": testcases
         }
-        self.runner.run_tests(tests_mapping)
-        summary = self.runner.summary
+        summary = self.runner.run_tests(tests_mapping)
         self.assertTrue(summary["success"])
 
     def test_run_httprunner_with_teardown_hooks_not_exist_attribute(self):
@@ -366,8 +364,7 @@ class TestHttpRunner(ApiServerUnittest):
             "project_mapping": loader.project_mapping,
             "testcases": testcases
         }
-        self.runner.run_tests(tests_mapping)
-        summary = self.runner.summary
+        summary = self.runner.run_tests(tests_mapping)
         self.assertFalse(summary["success"])
         self.assertEqual(summary["stat"]["teststeps"]["errors"], 1)
 
@@ -397,15 +394,13 @@ class TestHttpRunner(ApiServerUnittest):
             "project_mapping": loader.project_mapping,
             "testcases": testcases
         }
-        self.runner.run_tests(tests_mapping)
-        summary = self.runner.summary
+        summary = self.runner.run_tests(tests_mapping)
         self.assertFalse(summary["success"])
         self.assertEqual(summary["stat"]["teststeps"]["errors"], 1)
 
     def test_run_api(self):
         path = "tests/httpbin/api/get_headers.yml"
-        self.runner.run(path)
-        summary = self.runner.summary
+        summary = self.runner.run(path)
         self.assertTrue(summary["success"])
         self.assertEqual(summary["stat"]["testcases"]["total"], 1)
         self.assertEqual(summary["stat"]["teststeps"]["total"], 1)
@@ -413,8 +408,7 @@ class TestHttpRunner(ApiServerUnittest):
 
     def test_request_302_logs(self):
         path = "tests/httpbin/api/302_redirect.yml"
-        self.runner.run(path)
-        summary = self.runner.summary
+        summary = self.runner.run(path)
         self.assertTrue(summary["success"])
         self.assertEqual(summary["stat"]["testcases"]["total"], 1)
         self.assertEqual(summary["stat"]["teststeps"]["total"], 1)
@@ -425,10 +419,22 @@ class TestHttpRunner(ApiServerUnittest):
         self.assertEqual(req_resp_data[0]["response"]["status_code"], 302)
         self.assertEqual(req_resp_data[1]["response"]["status_code"], 200)
 
+    def test_request_302_logs_teardown_hook(self):
+        path = "tests/httpbin/api/302_redirect_teardown_hook.yml"
+        summary = self.runner.run(path)
+        self.assertTrue(summary["success"])
+        self.assertEqual(summary["stat"]["testcases"]["total"], 1)
+        self.assertEqual(summary["stat"]["teststeps"]["total"], 1)
+        self.assertEqual(summary["stat"]["teststeps"]["successes"], 1)
+
+        req_resp_data = summary["details"][0]["records"][0]["meta_datas"]["data"]
+        self.assertEqual(len(req_resp_data), 2)
+        self.assertEqual(req_resp_data[0]["response"]["status_code"], 302)
+        self.assertEqual(req_resp_data[1]["response"]["status_code"], 500)
+
     def test_request_with_params(self):
         path = "tests/httpbin/api/302_redirect.yml"
-        self.runner.run(path)
-        summary = self.runner.summary
+        summary = self.runner.run(path)
         self.assertTrue(summary["success"])
         self.assertEqual(summary["stat"]["testcases"]["total"], 1)
         self.assertEqual(summary["stat"]["teststeps"]["total"], 1)
@@ -443,20 +449,22 @@ class TestHttpRunner(ApiServerUnittest):
 
     def test_run_api_folder(self):
         api_folder = "tests/httpbin/api/"
-        self.runner.run(api_folder)
-        summary = self.runner.summary
+        summary = self.runner.run(api_folder)
+        print(summary["stat"]["testcases"]["total"])
+        print(len(summary["details"]))
         self.assertTrue(summary["success"])
-        self.assertEqual(summary["stat"]["testcases"]["total"], 2)
-        self.assertEqual(summary["stat"]["teststeps"]["total"], 2)
-        self.assertEqual(summary["stat"]["teststeps"]["successes"], 2)
-        self.assertEqual(len(summary["details"]), 2)
+        self.assertEqual(summary["stat"]["testcases"]["total"], 3)
+        self.assertEqual(summary["stat"]["teststeps"]["total"], 3)
+        self.assertEqual(summary["stat"]["teststeps"]["successes"], 3)
+        self.assertEqual(len(summary["details"]), 3)
         self.assertEqual(summary["details"][0]["stat"]["total"], 1)
         self.assertEqual(summary["details"][1]["stat"]["total"], 1)
+        self.assertEqual(summary["details"][2]["stat"]["total"], 1)
+
 
     def test_run_testcase_hardcode(self):
         for testcase_file_path in self.testcase_file_path_list:
-            self.runner.run(testcase_file_path)
-            summary = self.runner.summary
+            summary = self.runner.run(testcase_file_path)
             self.assertTrue(summary["success"])
             self.assertEqual(summary["stat"]["testcases"]["total"], 1)
             self.assertEqual(summary["stat"]["teststeps"]["total"], 3)
@@ -466,30 +474,26 @@ class TestHttpRunner(ApiServerUnittest):
     def test_run_testcase_template_variables(self):
         testcase_file_path = os.path.join(
             os.getcwd(), 'tests/data/demo_testcase_variables.yml')
-        self.runner.run(testcase_file_path)
-        summary = self.runner.summary
+        summary = self.runner.run(testcase_file_path)
         self.assertTrue(summary["success"])
 
     def test_run_testcase_template_import_functions(self):
         testcase_file_path = os.path.join(
             os.getcwd(), 'tests/data/demo_testcase_functions.yml')
-        self.runner.run(testcase_file_path)
-        summary = self.runner.summary
+        summary = self.runner.run(testcase_file_path)
         self.assertTrue(summary["success"])
 
     def test_run_testcase_layered(self):
         testcase_file_path = os.path.join(
             os.getcwd(), 'tests/data/demo_testcase_layer.yml')
-        self.runner.run(testcase_file_path)
-        summary = self.runner.summary
+        summary = self.runner.run(testcase_file_path)
         self.assertTrue(summary["success"])
         self.assertEqual(len(summary["details"]), 1)
 
     def test_run_testcase_output(self):
         testcase_file_path = os.path.join(
             os.getcwd(), 'tests/data/demo_testcase_layer.yml')
-        self.runner.run(testcase_file_path)
-        summary = self.runner.summary
+        summary = self.runner.run(testcase_file_path)
         self.assertTrue(summary["success"])
         self.assertIn("token", summary["details"][0]["in_out"]["out"])
         # TODO: add
@@ -501,8 +505,7 @@ class TestHttpRunner(ApiServerUnittest):
         variables_mapping = {
             "app_version": '2.9.7'
         }
-        self.runner.run(testcase_file_path, mapping=variables_mapping)
-        summary = self.runner.summary
+        summary = self.runner.run(testcase_file_path, mapping=variables_mapping)
         self.assertTrue(summary["success"])
         self.assertIn("token", summary["details"][0]["in_out"]["out"])
         # TODO: add
@@ -511,8 +514,7 @@ class TestHttpRunner(ApiServerUnittest):
     def test_run_testcase_with_parameters(self):
         testcase_file_path = os.path.join(
             os.getcwd(), 'tests/testsuites/create_users_with_parameters.yml')
-        self.runner.run(testcase_file_path)
-        summary = self.runner.summary
+        summary = self.runner.run(testcase_file_path)
         self.assertTrue(summary["success"])
         self.assertEqual(len(summary["details"]), 3 * 2)
 
@@ -550,8 +552,8 @@ class TestHttpRunner(ApiServerUnittest):
     #     # TODO: fix compatibility with Python 2.7
     #     testcase_file_path = os.path.join(
     #         os.getcwd(), 'tests/httpbin/basic.yml')
-    #     self.runner.run(testcase_file_path)
-    #     self.assertTrue(self.runner.summary["success"])
+    #     summary = self.runner.run(testcase_file_path)
+    #     self.assertTrue(summary["success"])
 
     def test_html_report_xss(self):
         testcases = [
@@ -583,7 +585,8 @@ class TestHttpRunner(ApiServerUnittest):
         tests_mapping = {
             "testcases": testcases
         }
-        report_path = self.runner.run(tests_mapping)
+        summary = self.runner.run(tests_mapping)
+        report_path = report.gen_html_report(summary)
         with open(report_path) as f:
             content = f.read()
             m = re.findall(
@@ -788,18 +791,3 @@ class TestApi(ApiServerUnittest):
             results.records[1]["name"],
             "create user and check result."
         )
-
-
-class TestLocust(unittest.TestCase):
-
-    def test_prepare_locust_tests(self):
-        path = os.path.join(
-            os.getcwd(), 'tests/locust_tests/demo_locusts.yml')
-        locust_tests = prepare_locust_tests(path)
-        self.assertEqual(len(locust_tests), 2 + 3)
-        name_list = [
-            "create user 1000 and check result.",
-            "create user 1001 and check result."
-        ]
-        self.assertIn(locust_tests[0]["config"]["name"], name_list)
-        self.assertIn(locust_tests[4]["config"]["name"], name_list)